rec/requirements.txt
torch>=2.0.0
transformers>=4.30.0
faiss-cpu>=1.7.4
fastapi>=0.100.0
uvicorn>=0.23.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
lightgbm>=4.0.0
pytest>=7.4.0
httpx>=0.24.0
